updates on equal height of blog content. Delete Modal incorporated also
I have worked on the blog page, whereby the blog content now have equal height irrespective of their content.
Also the delete popup that will be displayed when a user want to delete some content on the dashboard has be created. (check it under the view admin page - ViewAdmin.vue on Admin dashboard)